+++ tests/bin/run_elasticsearch.sh
@@ -33,6 +33,9 @@ do
     echo "http.port: $http_port" >> $config_yml
     echo "thrift.port: $thrift_port" >> $config_yml
 
+    # enable dynamic scripting
+    echo "script.disable_dynamic: false" >> $config_yml
+
     # enable udp
     echo "bulk.udp.enabled: true" >> $config_yml
     echo "bulk.udp.bulk_actions: 5" >> $config_yml