+++ tests/facet/test_termsstats.py
@@ -2,7 +2,7 @@ __author__ = 'Joe Linn'
 
 import unittest
 import pylastica
-from ..base import Base
+from tests.base import Base
 
 
 class TermsStatsTest(unittest.TestCase, Base):