+++ changes.markdown
@@ -1,5 +1,8 @@
 Changelog
 =========
+### 2014-06-19
+* Update to Elasticsearch 1.2.1
+
 ### 2014-04-02
 * Add `set_size()` and `set_shard_size()` methods to [Terms aggregation](https://github.com/jlinn/pylastica/blob/master/pylastica/aggregation/terms.py)